refactor TxData
- separate into specified components `HexEncodedData`, `MethodDetails` & `MultiSendDetails`
